requirements.txt
bleach==6.2.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
defusedxml==0.7.1
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
python-dotenv==1.1.0
requests==2.32.4
urllib3==2.4.0
webencodings==0.5.1
Werkzeug==3.1.3
youtube-transcript-api==1.1.0
